// src/enums/wendeType.ts
// src/enums/wendeType.ts
export enum WendeType{
  ZomerZonnewende = 'zomerzonnewende',
  WinterZonnewende = 'winterzonnewende',
  NoordGroteMaanwende = 'noordelijke grote maanwende',
  NoordKleineMaanwende = 'noordelijke kleine maanwende',
  ZuidGroteMaanwende = 'zuidelijke grote maanwende',
  ZuidKleineMaanwende = 'zuidelijke kleine maanwende',
}